--- packages/social-connection/README.md
@@ -1,29 +1,0 @@
-# Social Connection DataModel
-
-## Installation
-
-```sh
-npm install -D @datamodels/social-connection
-```
-
-## Rationale
-
-People are getting tired of registering and re-declaring their friend relationships on every site. This is especially true in web3 where each user's identity is already portable across sites.
-
-With a decentralized social connection owned by users instead of platforms, each user is able to leave an app and join a new one without losing any connections or having to migrate everyone.
-
-New social applications could all focus on creating and curating the best content and forget about rebuilding the wheels of a social connection database.
-
-The **Social Connection** contains a DID's basic social connection information.
-
-By standardizing connection information for DIDs, the Social Connection simplifies how applications can view and display the connections of their users.
-
-## Schemas
-
-### [SocialConnection](./schemas/SocialConnection.json)
-
-## [Discussion](https://github.com/ceramicstudio/datamodels/discussions/15)
-
-## License
-
-Dual licensed under MIT and Apache 2
