+++ packages/social-connections/README.md
@@ -1,0 +1,29 @@
+# Social Connections DataModel
+
+## Installation
+
+```sh
+npm install -D @datamodels/social-connections
+```
+
+## Rationale
+
+People are getting tired of registering and re-declaring their friend relationships on every site. This is especially true in web3 where each user's identity is already portable across sites.
+
+With a decentralized social connections owned by users instead of platforms, each user is able to leave an app and join a new one without losing any connections or having to migrate everyone.
+
+New social applications could all focus on creating and curating the best content and forget about rebuilding the wheels of a social connections database.
+
+The **Social Connections** contains a DID's basic social connections information.
+
+By standardizing connections information for DIDs, the Social Connections simplifies how applications can view and display the connections of their users.
+
+## Schemas
+
+### [SocialConnections](./schemas/SocialConnections.json)
+
+## [Discussion](https://github.com/ceramicstudio/datamodels/discussions/15)
+
+## License
+
+Dual licensed under MIT and Apache 2
