+++ packages/social-connections/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@datamodels/social-connection",
+  "name": "@datamodels/social-connections",
   "version": "0.1.3",
   "author": "CyberConnect",
   "license": "(Apache-2.0 OR MIT)",
@@ -11,7 +11,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/ceramicstudio/datamodels",
-    "directory": "packages/social-connection"
+    "directory": "packages/social-connections"
   },
   "main": "dist/model.cjs",
   "module": "dist/model.mjs",
