+++ packages/social-connections/prettier.config.js
