+++ packages/social-connections/schemas/SocialConnections.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema",
-  "title": "SocialConnection",
+  "title": "SocialConnections",
   "type": "object",
   "$defs": {
     "connection": {
