+++ packages/social-connections/tsconfig.json
